Moved test and options to correct places
